Drop table if exist microdonation sms logs

+++ db/migrate/20160506003558_drop_table_microdonation_sms_logs.rb
@@ -1,5 +1,5 @@
 class DropTableMicrodonationSmsLogs < ActiveRecord::Migration
   def change
-    drop_table :microdonation_sms_logs
+    drop_table :microdonation_sms_logs if ActiveRecord::Base.connection.table_exists? 'microdonation_sms_logs'
   end
 end
